movie-service/requirements.txt
asyncpg==0.21.0
bcrypt==3.2.0
cryptography==3.3.1
databases[postgresql]==0.4.1
fastapi==0.63.0
passlib==1.7.4
python-jose==3.2.0
python-multipart==0.0.5
SQLAlchemy==1.3.23
uvicorn==0.13.3
httpx==0.11.1

